// comonl-comonlwcl/src/app/modules/comonlapi/model/ruolo.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { TipoSoggettoAbilitato } from './tipoSoggettoAbilitato';


export interface Ruolo { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    denominazioneAzienda?: string;
    ilRuolo?: string;
    codiceFiscaleUtente?: string;
    codiceFiscaleAzienda?: string;
    dsNome?: string;
    dsCognome?: string;
    listaCasiUso?: Array<string>;
    codMinSoggettoAbilitato?: string;
    idSoggetti?: number;
    flgScuolaPubblica?: boolean;
    codiceFiscaleStudioProfessionale?: string;
    descrizioneStudioProfessionale?: string;
    flgConsulenteAccentrato?: boolean;
    delegaValida?: boolean;
    tipoAnagrafica?: string;
    email?: string;
    codiceEnteOperatoreAPL?: string;
    tipoSoggettoAbilitato?: TipoSoggettoAbilitato;
    consulenteRespo?: boolean;
    amministratore?: boolean;
    delegatoRespo?: boolean;
    legaleRappresentante?: boolean;
    personaAutorizzata?: boolean;
    operatoreProvinciale?: boolean;
    monitoraggio?: boolean;
}