// comonl-comonlwcl/src/app/modules/comonlapi/model/comunicazione.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { Cpi } from './cpi';
import { Datore } from './datore';
import { Delegato } from './delegato';
import { Lavoratore } from './lavoratore';
import { Provincia } from './provincia';
import { RapportiLavoratoriSediInteressateVD } from './rapportiLavoratoriSediInteressateVD';
import { Rapporto } from './rapporto';
import { StatoComunicazione } from './statoComunicazione';
import { TipoAcquisizione } from './tipoAcquisizione';
import { TipoComunicazione } from './tipoComunicazione';
import { TipoComunicazioneTu } from './tipoComunicazioneTu';
import { TipoProvenienza } from './tipoProvenienza';
import { TipoSoggettoAbilitato } from './tipoSoggettoAbilitato';
import { TipoSomministrazione } from './tipoSomministrazione';
import { TipoTracciato } from './tipoTracciato';
import { TipoTrasferimento } from './tipoTrasferimento';


export interface Comunicazione { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    idUserInsert?: string;
    idUserUltMod?: string;
    dtInsert?: Date;
    dtUltMod?: Date;
    annoDataOraRegionale?: string;
    annoProtCom?: number;
    annoProtMsgPosta?: number;
    annoProtRettifica?: number;
    annoProtUrgenza?: number;
    cfImpresa?: string;
    cfLavCAssunzione?: string;
    codiceComunRegPrec?: string;
    codiceComunicazioneReg?: string;
    codiceEnte?: string;
    codiceFiscaleSoggetto?: string;
    dsCausaForzamaggiore?: string;
    dsMotivoUrgenza?: string;
    dtAssCAssunzione?: Date;
    dtFineAffittoRamo?: Date;
    dtInvio?: Date;
    dtInvioMinistero?: Date;
    dtProtocollo?: Date;
    dtTimbroPostale?: Date;
    dtTrasferimentoVarDatori?: Date;
    dtVerifica?: Date;
    email?: string;
    emailsoggetto?: string;
    flgCausaForzamaggiore?: string;
    flgComDatLav?: string;
    flgComunSegUrgenza?: string;
    flgInvioMinistero?: string;
    flgMultiLav?: string;
    flgRettifica?: string;
    flgSommin?: string;
    idComDComunicPrecedente?: number;
    numAgenzSomministr?: number;
    numProtCom?: number;
    numProtComRettifica?: number;
    numProtComUrgenza?: number;
    numProtMsgPosta?: number;
    verifica?: string;
    idComDComunicazGruppo?: number;
    idComTuPrecedenteAnnullo?: number;
    idComTuDaRettificare?: number;
    delegato?: Delegato;
    cpi?: Cpi;
    provincia?: Provincia;
    statoComunicazione?: StatoComunicazione;
    tipoAcquisizione?: TipoAcquisizione;
    tipoComunicazione?: TipoComunicazione;
    tipoComunicazioneTu?: TipoComunicazioneTu;
    tipoProvenienza?: TipoProvenienza;
    tipoSoggettoAbilitato?: TipoSoggettoAbilitato;
    tipoSomministrazione?: TipoSomministrazione;
    tipoTracciato?: TipoTracciato;
    tipoTrasferimento?: TipoTrasferimento;
    dataRiferimento?: Date;
    idRapportoLavoro?: number;
    flgCurrentRecord?: string;
    datoreAttuale?: Datore;
    datorePrecedente?: Datore;
    lavoratori?: Array<Lavoratore>;
    lavoratoreCoObbligato?: Lavoratore;
    rapporto?: Rapporto;
    missione?: Rapporto;
    rapLavSedeVD?: Array<RapportiLavoratoriSediInteressateVD>;
    dtProtocolloPrec?: Date;
    annoProtComPrec?: number;
    numProtComPrec?: number;
    provinciaPrec?: Provincia;
}