// comonl-comonlwcl/src/app/modules/comonlapi/model/provincia.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { Regione } from './regione';


export interface Provincia { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    codProvinciaMin?: string;
    codiceFiscalePv?: string;
    dsComTProvincia?: string;
    dsTarga?: string;
    dtFine?: Date;
    dtInizio?: Date;
    dtTmst?: Date;
    regione?: Regione;
}