// comonl-comonlwcl/src/app/modules/comonlapi/model/wrapperComunicazione.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Comunicazione } from './comunicazione';
import { Ruolo } from './ruolo';


export interface WrapperComunicazione { 
    ruolo?: Ruolo;
    comunicazione?: Comunicazione;
}